pyproject.toml
[project]
name = "agenticrag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.1.0",
    "flask>=3.1.2",
    "groq>=0.31.1",
    "langchain-chroma>=0.2.6",
    "langchain-community>=0.3.29",
    "langchain-groq>=0.3.8",
    "langchain-huggingface>=0.3.1",
    "langchain[all]>=0.3.27",
    "langgraph>=0.6.7",
    "pinecone-client>=6.0.0",
    "python-dotenv>=1.1.1",
    "sentence-transformers>=5.1.0",
    "setuptools>=80.9.0",
    "streamlit>=1.49.1",
    "tavily-python>=0.7.12",
    "tqdm>=4.67.1",
]
